% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{accu}
\alias{accu}
\title{forecast accuracy}
\usage{
accu(actual, forecast, outlist = "mean,FBias,WMAPE")
}
\arguments{
\item{actual}{target column to predict}

\item{forecast}{prediction column}

\item{outlist}{export column. available options: n,mean,sd,CV,R2,DB,FBias,MPE,MAPE,RMSE,MAD,MADP,MASE,RAE,WMAPE}
}
\description{
measure critical metrics for forecasting
}
\examples{
 accu(actual=105,forecast=100,outlist="mean,WMAPE,FBias")
}
