% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{date_mdy_ymd}
\alias{date_mdy_ymd}
\title{change date format}
\usage{
date_mdy_ymd(x)
}
\arguments{
\item{x}{date value}
}
\description{
find latest sunday date
}
\examples{
exampleDataR= data.table::as.data.table(exampleDataR)
exampleDataR = exampleDataR[order(location,item,date)]
exampleDataR[,date_better:= date_mdy_ymd(date)]
head(exampleDataR,15)
}
