% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{date_to_yyyymmdd}
\alias{date_to_yyyymmdd}
\title{Convert date string to YYYY-MM-DD format}
\usage{
date_to_yyyymmdd(date_str, smart_format = T, forceformat = "\%Y.\%m.\%d")
}
\arguments{
\item{date_str}{A character string representing a date in an unknown format.}

\item{smart_format}{if it is T, then it will try to detect format. otherwise force format}

\item{forceformat}{to make this effective, smart_format should be False c("%Y-%m-%d", "%m/%d/%Y", "%d/%m/%Y", "%m-%d-%Y", "%d-%m-%Y", "%Y/%m/%d", "%Y.%m.%d", "%d.%m.%Y", "%Y%m%d", "%b %d, %Y")}
}
\value{
A character string representing the date in the YYYY-MM-DD format, or
        NA if the date string could not be converted.
}
\description{
This function attempts to detect the format of a date string and convert it to
the YYYY-MM-DD format. It supports a variety of commonly used date formats.
}
\examples{
date_to_yyyymmdd("2022-02-28")
date_to_yyyymmdd("02/28/2022")
date_to_yyyymmdd("28.02.2022")
date_to_yyyymmdd("Feb 28, 2022")

}
