% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataR}
\alias{exampleDataR}
\title{Sales data for two locations}
\format{
A data frame with 60 rows and 4 columns, 2 location 2 item:
\describe{
  \item{location}{city of the location, istanbul or ankara}
  \item{item}{sales of two items, coffee or tea}
  \item{date}{sales dates. 30 days}
  \item{sales}{random sales figures between 0 to 40}
}
}
\source{
https://github.com/mkparkin/Rinvent
}
\usage{
exampleDataR
}
\description{
location, item, date level sales data
}
\keyword{datasets}
