% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{fast.date}
\alias{fast.date}
\title{fast date conversion}
\usage{
fast.date(x)
}
\arguments{
\item{x}{date column to convert}
}
\description{
works faster than as.Date.
}
\examples{
 fast.date('as.factor('2021-01-01')')
}
