% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{faster_date}
\alias{faster_date}
\title{faster date conversion}
\usage{
faster_date(dt, dateCol = "Date")
}
\arguments{
\item{dt, dateCol}{}
}
\description{
works faster than fast.date
}
\examples{
 faster_date(dt,"Date")

}
