% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{lsos}
\alias{lsos}
\title{list variables and size}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{n}{top N rows}
}
\description{
list variables and size in r memory
}
\examples{
 data = data.table::as.data.table(exampleDataR)
lsos()
}
