% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_azure.R
\name{read_csv_azure}
\alias{read_csv_azure}
\title{read_csv_azure read csv files from azure blob}
\usage{
read_csv_azure(file_path, containerconnection)
}
\arguments{
\item{containerconnection}{blob connection}

\item{pathtoread}{blob path to read}
}
\description{
read_csv_azure read csv files from azure blob
}
\examples{
 read_csv_azure()
}
