% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readparquetR.R
\name{readparquetR}
\alias{readparquetR}
\title{reading parquet or delta files from local directory or aws s3 or azure blob}
\usage{
readparquetR(
  pathtoread,
  where = "",
  partition = "",
  collist = "",
  sample = F,
  samplesizecount = 3,
  add_part_names = F,
  format = "parquet",
  filelocation = "local",
  containerconnection = NULL,
  bucket = NULL
)
}
\arguments{
\item{pathtoread}{reading path, local or azure cloud}

\item{where}{it will read datatable and filter with this condition. i.e. you can write where="column='A'"}

\item{partition}{if you want to read a partition files with a pattern. i.e. partition=c('2017','2018')}

\item{collist}{specific columns to read}

\item{sample}{sample=T just to see sample rows. dont read whole table}

\item{samplesizecount}{default=3 rows. you can change it}

\item{add_part_names}{when it is partitioned, you need to make this T to add partition names as column}

\item{filelocation}{"local" or "azure" or "s3"}

\item{containerconnection}{if filelocation="azure" then we need connection. something like this}

\item{bucket}{if filelocation="s3" we need to put bucket name}
}
\value{
data table
}
\description{
reads parquet,delta files from local or cloud
}
\examples{
 sample_data <- readparquetR("localfolder/",
collist = c("column1"
           ,"column2"
           ,"column3"
),
format="delta",
where="SKU==19058901 & STORE!='1905'")
readparquetR(pathtoread="C:/users/...",add_part_names=F,collist="",sample=F,where="sku=1 & store=1",partition="2022")
}
