% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3save_with_tryCatch.R
\name{s3save_with_tryCatch}
\alias{s3save_with_tryCatch}
\title{save your files to Amazon S3 with "tryCatch" capability}
\usage{
s3save_with_tryCatch(df, format, object, bucket)
}
\arguments{
\item{df, }{format, object, bucket}
}
\description{
save your files with tryCatch() to implement retry capability
}
\examples{
s3save_with_tryCatch(df = results_df, format = "csv", object = file.path("target_s3_path.csv"), bucket = "bucket_name")

}
