% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{uniquen}
\alias{uniquen}
\title{count unique values in a table}
\usage{
uniquen(data)
}
\arguments{
\item{data}{input data table}
}
\description{
count unique values in a table
}
\examples{
 uniquen(exampleDataR)
}
