% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{week_start}
\alias{week_start}
\title{find latest sunday}
\usage{
week_start(date)
}
\arguments{
\item{date}{date value}
}
\description{
find latest sunday date
}
\examples{
exampleDataR= data.table::as.data.table(exampleDataR)
exampleDataR = exampleDataR[order(location,item,date)]
exampleDataR[,week_start_date:= week_start(date)]
head(exampleDataR,15)
}
