% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_blob.R
\name{write_to_blob}
\alias{write_to_blob}
\title{write_to_blob}
\usage{
write_to_blob(
  data,
  templocalpath = "default",
  savename = "my_df.parquet",
  containerconnection = NULL,
  blobfilepath = "trial",
  format = "parquet",
  overwrite = FALSE
)
}
\arguments{
\item{data}{A data object to write to Azure Blob storage}

\item{templocalpath}{The path to a temporary local directory to use for intermediate files (default is a new directory in the system temp directory)}

\item{savename}{The name to use for the file in Azure Blob storage}

\item{containerconnection}{The connection to the Azure Blob storage container}

\item{blobfilepath}{The path within the container to write the file (default is the root of the container)}

\item{format}{The format to use when writing the data (one of "parquet", "rds", or "csv")}

\item{overwrite}{overwrite existing file (one of "parquet", "rds", or "csv")}
}
\description{
Writes data to an Azure Blob storage container using the specified format
}
\details{
Write data to an Azure Blob storage container
}
\examples{
write_to_blob(datatable,
              savename = "data.parquet",
              containerconnection = mycontainer,
              blobfilepath = file.path("mainfolder", "subfolder"),
              overwrite=FALSE
              )

}
