% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_S3.R
\name{write_to_S3}
\alias{write_to_S3}
\title{save your files to Amazon S3 with "tryCatch" capability}
\usage{
write_to_S3(df, format, object, bucket)
}
\arguments{
\item{df, }{format, object, bucket}
}
\description{
save your files with tryCatch() to implement retry capability
}
\examples{
write_to_S3(df = results_df, format = "csv", object = file.path("target_s3_path.csv"), bucket = "bucket_name")

}
