% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{crossjoin}
\alias{crossjoin}
\title{table cross join}
\usage{
crossjoin(t1, t2)
}
\arguments{
\item{t1}{table 1}

\item{t2}{table 2}
}
\value{
crossjoin()
}
\description{
table cross join
}
\examples{
 crossjoin()
}
