% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{faster_date}
\alias{faster_date}
\title{faster date conversion}
\usage{
faster_date(dt, dateCol = "Date")
}
\arguments{
\item{dt}{# data table}

\item{dateCol}{# date column name}
}
\description{
works faster than fast.date
}
\examples{
 exampleDataR = data.table::as.data.table(exampleDataR)
exampleDataR = exampleDataR[,dateformat:= date_mdy_ymd(date)]
faster_date(exampleDataR,"dateformat")
}
